Move configuration validation from Common to Core.
