+++ src/backend/TrafficCourts/Citizen.Service/Services/Tickets/Search/Mock/FileMockDataProvider.cs
@@ -1,5 +1,4 @@
-﻿using TrafficCourts.Common.Configuration.Validation;
-using TrafficCourts.Configuration.Validation;
+﻿using TrafficCourts.Configuration.Validation;
 
 namespace TrafficCourts.Citizen.Service.Services.Tickets.Search.Mock
 {
