+++ src/backend/TrafficCourts/Common/OpenAPIs/OracleDataAPI/OracleDataApiConfiguration.cs
@@ -1,5 +1,4 @@
 ﻿using System.ComponentModel.DataAnnotations;
-using TrafficCourts.Common.Configuration.Validation;
 using TrafficCourts.Configuration.Validation;
 
 namespace TrafficCourts.Common.OpenAPIs.OracleDataAPI;
