+++ src/backend/TrafficCourts/TrafficCourts.Arc.Dispute.Service/Configuration/SftpOptions.cs
@@ -1,5 +1,4 @@
-﻿using TrafficCourts.Common.Configuration.Validation;
-using TrafficCourts.Configuration.Validation;
+﻿using TrafficCourts.Configuration.Validation;
 
 namespace TrafficCourts.Arc.Dispute.Service.Configuration
 {
