+++ src/backend/TrafficCourts/Workflow.Service/Configuration/EmailConfiguration.cs
@@ -1,5 +1,4 @@
-﻿using TrafficCourts.Common.Configuration.Validation;
-using TrafficCourts.Configuration.Validation;
+﻿using TrafficCourts.Configuration.Validation;
 
 namespace TrafficCourts.Workflow.Service.Configuration;
